sweet alerts done

+++ clientsrc/src/components/challenge.vue
@@ -72,6 +72,8 @@
 </template>
 
 <script>
+import Swal from "sweetalert2";
+
 export default {
   name: "Challenge",
   props: ["challengeData"],
@@ -95,7 +97,20 @@ export default {
       $("#modal" + this.challengeData._id).modal("toggle");
     },
     deleteChallenge(challenge) {
-      this.$store.dispatch("deleteChallengeById", challenge);
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          Swal.fire("Deleted!", "Challenge has been deleted.", "success");
+          this.$store.dispatch("deleteChallengeById", challenge);
+        }
+      });
     }
   },
   computed: {
