+++ clientsrc/src/components/task.vue
@@ -25,12 +25,27 @@
 </template>
 
 <script>
+import Swal from "sweetalert2";
+
 export default {
   name: "Task",
   props: ["taskData"],
   methods: {
     deleteTask(task) {
-      this.$store.dispatch("deleteTaskById", task);
+      Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+      }).then(result => {
+        if (result.value) {
+          Swal.fire("Deleted!", "Task has been deleted.", "success");
+          this.$store.dispatch("deleteTaskById", task);
+        }
+      });
     }
   }
 };
